docs-requirements.txt
sphinx
m2r2
furo
pypandoc_binary
sphinx-copybutton
